Move run into a subdirectory and split `iter` stuff into a module
